Automatically generate dataset group based on project slug and version
Right now the group is set via the `DataAccess` model which has to be
handled in the host application (PhysioNet). To fit all regexes checked
in the API, only alfanumerics are allowed in the group name.
Automatically generating the parameter this was might yield edge cases
in very obscure cases, but it's very likely to never occur.

Making this 100% robust can be done by introducing further
transformations on the API side (two parameters/replacing certain characters).